+++ docs/Integration-with-CxSCA.md
@@ -176,7 +176,7 @@ In order to open SCA security tickets, set the bug tracker in cxflow app.yml fil
 
 
 ## <a name="zipFolderScan">SCA ZIP folder scan</a>
-In order to change the default Cx-Flow SCA scan behaviour and to perform a SCA ZIP scan, the next configuration property should be added underneath the sca configuration section:
+In order to change the default CxFlow SCA scan behaviour and to perform a SCA ZIP scan, the next configuration property should be added underneath the sca configuration section:
 ```
 enabledZipScan: true
 ```
