+++ src/main/java/com/checkmarx/flow/service/ASTScanner.java
@@ -1,22 +1,13 @@
 package com.checkmarx.flow.service;
 
 import com.checkmarx.flow.config.FlowProperties;
-
 import com.checkmarx.sdk.config.AstProperties;
-
-import com.checkmarx.sdk.config.ScaProperties;
 import com.checkmarx.sdk.dto.ScanResults;
-
 import com.checkmarx.sdk.dto.ast.ASTResults;
 import com.checkmarx.sdk.dto.ast.ASTResultsWrapper;
-
-
-import com.checkmarx.sdk.service.CxRepoFileService;
 import com.cx.restclient.AstClientImpl;
-
 import com.cx.restclient.ast.dto.sast.AstSastResults;
 import lombok.extern.slf4j.Slf4j;
-
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -27,9 +18,8 @@ import java.util.Optional;
 @Slf4j
 public class ASTScanner extends AbstractASTScanner {
 
-    public ASTScanner(AstClientImpl astClient, FlowProperties flowProperties, BugTrackerEventTrigger bugTrackerEventTrigger,
-                      ScaProperties scaProperties, CxRepoFileService cxRepoFileService) {
-        super(astClient, flowProperties, AstProperties.CONFIG_PREFIX, bugTrackerEventTrigger, scaProperties, cxRepoFileService);
+    public ASTScanner(AstClientImpl astClient, FlowProperties flowProperties, BugTrackerEventTrigger bugTrackerEventTrigger) {
+        super(astClient, flowProperties, AstProperties.CONFIG_PREFIX, bugTrackerEventTrigger);
     }
 
     @Override
