+++ src/main/java/com/checkmarx/flow/service/AbstractASTScanner.java
@@ -9,19 +9,13 @@ import com.checkmarx.flow.dto.report.AnalyticsReport;
 import com.checkmarx.flow.dto.report.ScanReport;
 import com.checkmarx.flow.exception.ExitThrowable;
 import com.checkmarx.flow.exception.MachinaRuntimeException;
-import com.checkmarx.sdk.config.Constants;
-import com.checkmarx.sdk.config.ScaProperties;
 import com.checkmarx.sdk.dto.ScanResults;
 import com.checkmarx.sdk.dto.ast.ASTResults;
 import com.checkmarx.sdk.dto.ast.ASTResultsWrapper;
 import com.checkmarx.sdk.dto.ast.SCAResults;
 import com.checkmarx.sdk.dto.ast.ScanParams;
-import com.checkmarx.sdk.dto.cx.CxScanParams;
-import com.checkmarx.sdk.exception.CheckmarxException;
-import com.checkmarx.sdk.service.CxRepoFileService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.net.MalformedURLException;
@@ -36,21 +30,12 @@ public abstract class AbstractASTScanner implements VulnerabilityScanner {
     private final FlowProperties flowProperties;
     private final String scanType;
     private final BugTrackerEventTrigger bugTrackerEventTrigger;
-    private final ScaProperties scaProperties;
-    private final CxRepoFileService cxRepoFileService;
 
     @Override
     public ScanResults scan(ScanRequest scanRequest) {
         ScanResults result = null;
         log.info("--------------------- Initiating new {} scan ---------------------", scanType);
         ScanParams sdkScanParams = toSdkScanParams(scanRequest);
-
-        if (scaProperties.isEnabledZipScan()) {
-            log.info("CxAST-SCA zip scan is enabled");
-            String scaZipFolderPath = getScaZipFolderPath(scanRequest);
-            sdkScanParams.setZipPath(scaZipFolderPath);
-        }
-
         ASTResultsWrapper internalResults = new ASTResultsWrapper(new SCAResults(), new ASTResults());
         try {
             bugTrackerEventTrigger.triggerScanStartedEvent(scanRequest);
@@ -102,24 +87,8 @@ public abstract class AbstractASTScanner implements VulnerabilityScanner {
         return toScanResults(internalResults);
     }
 
-    private String getScaZipFolderPath(ScanRequest scanRequest) {
-        CxScanParams cxScanParams = prepareScanParamsToCloneRepo(scanRequest);
-        try {
-            return cxRepoFileService.prepareRepoFile(cxScanParams);
-        } catch (CheckmarxException e) {
-            throw new MachinaRuntimeException(e.getMessage());
-        }
-    }
-
-    private CxScanParams prepareScanParamsToCloneRepo(ScanRequest scanRequest) {
-        CxScanParams cxScanParams = new CxScanParams();
-        cxScanParams.withGitUrl(scanRequest.getRepoUrlWithAuth());
-        cxScanParams.withFileExclude(scanRequest.getExcludeFiles());
-
-        if (StringUtils.isNotEmpty(scanRequest.getBranch())) {
-            cxScanParams.withBranch(Constants.CX_BRANCH_PREFIX.concat(scanRequest.getBranch()));
-        }
-        return cxScanParams;
+    protected void setScannerSpecificProperties(ScanRequest scanRequest, ScanParams scanParams) {
+        log.info("No scanner specific properties were initialized.");
     }
 
     private void treatError(ScanRequest scanRequest, ASTResultsWrapper internalResults, Exception e) {
@@ -174,12 +143,15 @@ public abstract class AbstractASTScanner implements VulnerabilityScanner {
     private ScanParams toSdkScanParams(ScanRequest scanRequest) {
         URL parsedUrl = getRepoUrl(scanRequest);
 
-        return ScanParams.builder()
+        ScanParams scanParams = ScanParams.builder()
                 .projectName(scanRequest.getProject())
                 .remoteRepoUrl(parsedUrl)
                 .scaConfig(scanRequest.getScaConfig())
                 .filterConfiguration(scanRequest.getFilter())
                 .build();
+
+        setScannerSpecificProperties(scanRequest, scanParams);
+        return scanParams;
     }
 
     private URL getRepoUrl(ScanRequest scanRequest) {
